requirements.txt
aiohttp>=3.7.4
arrow>=1.3.0
GitPython>=3.1.42
halo @ git+https://github.com/guigoruiz1/halo.git
ipykernel>=6.29.3
ipylab>=1.0.0
ipython>=8.12.3
itables>=1.7.1
jupyter_client>=8.6.1
matplotlib>=3.8.3
matplotlib_inline>=0.1.2
matplotlib_venn>=0.11.10
nbconvert>=7.11.0
nbformat>=5.10.3
numpy>=1.26.4
pandas>=2.2.1
papermill>=2.5.0
platformdirs>=2.4.0
python-dotenv>=1.0.1
Requests>=2.31.0
seaborn>=0.13.2
sphinx>=7.2.6
sphinx_rtd_theme>=2.0.0
termcolor>=2.4.0
tqdm>=4.62.3
wikitextparser>=0.55.8
